Moved from geven to cherrypy.
Due to some issues with multiprocessing, greenlets, ssl, and Ansible we
decided to move off of gevent. This change replaces gevent with
cherrypy. While cherrypy does offer it's own framework this only
utilizes it for it's server and continues to use falcon for the REST
framework.

Pull request: #65
Approved by: mbarnes