bug: etcd.Client now times out after 2 seconds.
When the commissaire server starts it attempts to get some configuration
from etcd. If the etcd host/port hangs the server would hang. This adds
a 2 second timeout to the etcd.Client reads so that the server exits
on start up with a failure instead of hanging forver.

Closes #98.

Closes: #99
Approved by: mbarnes