authentication is now configurable via CLI switch.
The authentication systhem was designed to be pluggable. However, early
on we were hardcoding the plugins as a primary and backup. This adds the
ability to define a plugin to use as well as any configuration
parameters it may need. By default the
commissaire.authentication.httpauthbyetcd is used.

Closes: #100
Approved by: mbarnes