ContainerManagerConfig Model (#50)
* constants: Added CONTAINER_MANAGER constants.

* models: Added ContainerManagerConfig.

The new ContainerManagerConfig model is meant to store configuration
data used when creating the upcomming ContainerManager instance. When a
cluster has a ContainerManager associated with it the controlling code
should match the Cluster.container_manager to a
ContainerManagerConfig.name. The resulting ContainerManager can then be
used to work with proper external ContainerManager.

Note that not all Clusters need a ContainerManager. The Cluster is not
associated with a ContainerManager if it's container_manager field is
empty.

* test: Added Model._coerce tests.

* doc: Added endpoints and enums for ContainerManagerConfig.

* models: Replaced Cluster.type with Cluster.container_manager.

Instead of binding a cluster to a single type, a series of fields will
be used to define cluster features. The first of the features being
container_manager which relates to the name of the
ContainerManagerConfig a Cluster is associated with.

* models: Added ContainerManagerConfigs.

* fixup! models: Replaced Cluster.type with Cluster.container_manager.

* fixup! doc: Added endpoints and enums for ContainerManagerConfig.