vagrant: Services now listen on all interfaces.
When specifying an interface for services to listen some services throw
errors noting they are unable to use the address. Upon manually starting
the service post vagrant up the service works fine. This change switches
to service binding to 0.0.0.0 which allows the service to come up and
listen on it's external (as well as any other given) interface(s).

This change also sets autostart to false for the commissaire machine to
match the kubernetes box.