util: Simplify logging configuration.
This simplifies logging configuration, both in the file format and
internally.

For starters, we return to the logging.config schema in the JSON
configuration file, under the "logging" key.  But this should only
be necessary for changing specific loggers or log message handling.

For convenience, DEBUG level log messages can be enabled for all
loggers by setting a "debug" option to True in the configuration
file.

Internally, rewrite setup_logging() to install a default handler
for the root logger only, instead of for each and every non-root
logger.  Non-root loggers will most likely propagate messages to
the root logger, unless the configuration overrides this.

This also eliminates the maintenance cost of having to specify
all logger names when calling setup_logging().