storage: Add NotifyCallback decorator.
Decorator for notification callback methods.

This is for methods passed to StorageClient.register_callback().

The modifed call signature for notification callbacks with this
decorator is (event, model, message), where 'event' is guaranteed to be
one of 'created', 'deleted' or 'updated'; and 'model' is an instance of
some class from commissaire.Models.