e2e: Update for Custodia integration.
Yet another service to start...

As with the other services, a private Custodia instance can be started
for the duration of the test suite with:

   behave -D start-custodia

or

   behave -D start-all-servers

Also, the step "a host already exists at {host}" requires talking
directly to Custodia's REST API instead of through Commissaire.  This in
turn requires a special requests.Session with UNIX Domain Socket support
as in CustodiaStoreHandler.