Don't require commctl in test environments.
commctl still depends on py-bcrypt, but commissaire-http recently
switched back to bcrypt.  The problem is, both packages install a
module named "bcrypt" and when py-bcrypt is installed it tends to
win in "import bcrypt" calls.  The two APIs are not compatible.

The test environments don't do anything with commctl, so leave it
out.